scripts/pyproject.toml
[tool.poetry]
name = "howfar"
version = "0.1.0"
description = ""
authors = ["ahmed <ahmed-machine@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
osmium = "^4.2.0"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
